+++ hutool-core/src/main/java/cn/hutool/core/lang/ClassScaner.java
@@ -81,7 +81,7 @@ public class ClassScaner {
 	}
 
 	/**
-	 * 扫面该包路径下所有class文件
+	 * 扫描该包路径下所有class文件
 	 * 
 	 * @return 类集合
 	 */
@@ -90,7 +90,7 @@ public class ClassScaner {
 	}
 
 	/**
-	 * 扫面该包路径下所有class文件
+	 * 扫描该包路径下所有class文件
 	 * 
 	 * @param packageName 包路径 com | com. | com.abs | com.abs.
 	 * @return 类集合
@@ -100,7 +100,7 @@ public class ClassScaner {
 	}
 
 	/**
-	 * 扫面包路径下满足class过滤器条件的所有class文件，<br>
+	 * 扫描包路径下满足class过滤器条件的所有class文件，<br>
 	 * 如果包路径为 com.abs + A.class 但是输入 abs会产生classNotFoundException<br>
 	 * 因为className 应该为 com.abs.A 现在却成为abs.A,此工具类对该异常进行忽略处理<br>
 	 * 
@@ -156,7 +156,7 @@ public class ClassScaner {
 	}
 
 	/**
-	 * 扫面包路径下满足class过滤器条件的所有class文件
+	 * 扫描包路径下满足class过滤器条件的所有class文件
 	 * 
 	 * @return 类集合
 	 */