+++ hutool-core/src/main/java/cn/hutool/core/lang/func/Func0.java
@@ -16,6 +16,7 @@ public interface Func0<R> {
 	 * 执行函数
 	 * 
 	 * @return 函数执行结果
+	 * @throws Exception 自定义异常
 	 */
-	R call();
+	R call() throws Exception;
 }