+++ hutool-db/src/test/java/cn/hutool/db/MySQLTest.java
@@ -6,10 +6,7 @@ import org.junit.Ignore;
 import org.junit.Test;
 
 import cn.hutool.core.lang.Console;
-import cn.hutool.db.Db;
-import cn.hutool.db.Entity;
-import cn.hutool.db.Page;
-import cn.hutool.db.PageResult;
+import cn.hutool.core.lang.func.VoidFunc1;
 
 /**
  * MySQL操作单元测试
@@ -30,6 +27,30 @@ public class MySQLTest {
 					.set("test1", "t" + id)//
 			);
 		}
+	}
+	
+	/**
+	 * 事务测试<br>
+	 * 更新三条信息，低2条后抛出异常，正常情况下三条都应该不变
+	 * 
+	 * @throws SQLException
+	 */
+	@Test(expected=SQLException.class)
+	@Ignore
+	public void txTest() throws SQLException {
+		Db.use("mysql").tx(new VoidFunc1<Db>() {
+			
+			@Override
+			public void call(Db db) throws Exception {
+				int update = db.update(Entity.create("user").set("text", "描述100"), Entity.create().set("id", 100));
+				db.update(Entity.create("user").set("text", "描述101"), Entity.create().set("id", 101));
+				if(1 == update) {
+					// 手动指定异常，然后测试回滚触发
+					throw new RuntimeException("Error");
+				}
+				db.update(Entity.create("user").set("text", "描述102"), Entity.create().set("id", 102));
+			}
+		});
 	}
 
 	@Test