+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/enjoy/EnjoyEngine.java
@@ -4,6 +4,7 @@ import org.beetl.core.GroupTemplate;
 
 import com.jfinal.template.source.FileSourceFactory;
 
+import cn.hutool.core.io.FileUtil;
 import cn.hutool.core.lang.Assert;
 import cn.hutool.core.util.ObjectUtil;
 import cn.hutool.extra.template.Template;
@@ -67,7 +68,7 @@ public class EnjoyEngine implements TemplateEngine {
 	private static com.jfinal.template.Engine createEngine(TemplateConfig config) {
 		Assert.notNull(config, "Template config is null !");
 		final com.jfinal.template.Engine engine = com.jfinal.template.Engine.create("Hutool-Enjoy-Engine-" + config.toString());
-		engine.setEncoding(config.getCharset().toString());
+		engine.setEncoding(config.getCharsetStr());
 
 		switch (config.getResourceMode()) {
 		case STRING:
@@ -79,6 +80,11 @@ public class EnjoyEngine implements TemplateEngine {
 			break;
 		case FILE:
 			engine.setSourceFactory(new FileSourceFactory());
+			engine.setBaseTemplatePath(config.getPath());
+			break;
+		case WEB_ROOT:
+			engine.setSourceFactory(new FileSourceFactory());
+			engine.setBaseTemplatePath(FileUtil.getAbsolutePath(FileUtil.getWebRoot()));
 			break;
 		default:
 			break;