+++ hutool-extra/src/main/java/cn/hutool/extra/template/engine/velocity/SimpleStringResourceLoader.java
@@ -1,11 +1,13 @@
 package cn.hutool.extra.template.engine.velocity;
 
 import java.io.InputStream;
+import java.io.Reader;
+import java.io.StringReader;
 
-import org.apache.commons.collections.ExtendedProperties;
 import org.apache.velocity.exception.ResourceNotFoundException;
 import org.apache.velocity.runtime.resource.Resource;
 import org.apache.velocity.runtime.resource.loader.ResourceLoader;
+import org.apache.velocity.util.ExtProperties;
 
 import cn.hutool.core.io.IoUtil;
 import cn.hutool.core.util.CharsetUtil;
@@ -17,15 +19,26 @@ import cn.hutool.core.util.CharsetUtil;
  * @author looly
  *
  */
-public class SimpleStringResourceLoader extends ResourceLoader{
+public class SimpleStringResourceLoader extends ResourceLoader {
 
 	@Override
-	public void init(ExtendedProperties configuration) {
+	public void init(ExtProperties configuration) {
+	}
+
+	/**
+	 * 获取资源流
+	 * 
+	 * @param source 字符串模板
+	 * @return 流
+	 * @throws ResourceNotFoundException 资源未找到
+	 */
+	public InputStream getResourceStream(String source) throws ResourceNotFoundException {
+		return IoUtil.toStream(source, CharsetUtil.CHARSET_UTF_8);
 	}
 	
 	@Override
-	public InputStream getResourceStream(String source) throws ResourceNotFoundException {
-		return IoUtil.toStream(source, CharsetUtil.CHARSET_UTF_8);
+	public Reader getResourceReader(String source, String encoding) throws ResourceNotFoundException {
+		return new StringReader(source);
 	}
 
 	@Override